// nest_cardekho_app/src/used_car_details/dto/create-used_car_detail.dto.ts
export class CreateUsedCarDetailDto {
    name: string;
    title: string;
    images: string;
    carNumber: string;
    price: string;
    lenght: string;
    state: string;
    city: string;
    fuelType: string;
    body_type: string;
    registrationYear: string;
    insurance: string;
    seats: number;
    kmsDriven: string;
    Rto: string;
    ownership: string;
    engine: string;
    mileage: string;
    power: string;
    numberofAirbags: number;
    engineDisplacement: string;
    transmission: string;
    yearofManufacture: number;
    categoryId: number;
    sellerId: number;
    makeId:number;
}